accounts/urls.py
from rest_framework_simplejwt.views import (
    TokenRefreshView,
    TokenVerifyView,
    TokenObtainPairView,
)
from django.urls import path
from .views import RegisterAPIView,ProfileAPIView

urlpatterns = [
    path("auth/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("auth/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("register/", RegisterAPIView.as_view(), name="register"),  # for registering users
    path("profile/", ProfileAPIView.as_view(), name="profile"),
]